Create auto-tent.ino
Turns on board LED off and on with status of motion detection. Prints Serial 'Turn LEDs ON' and 'Turn LEDs OFF' when LEDs should be turned off and on according to the noMotionPeriod time set (miliseconds). It also accounts for 3 seconds of non-motion sensing timeout that happens when the HC-SR501 changes from HIGH to LOW.